# libs/sdk/swagger.json
{"openapi":"3.2.0","servers":[{"url":"http://localhost:5000/api/v1","description":"Local development"}],"info":{"version":"1.0.0","title":"@moringa/backend","description":"Moringa E-Commerce Backend API built with NestJS, Nestia, and Prisma"},"paths":{"/":{"get":{"tags":["app"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"additionalProperties":false,"properties":{"status":{"type":"string"},"app":{"type":"string"},"version":{"type":"string"}},"required":["status","app","version"],"type":"object"}}}}},"x-samchon-accessor":["root"],"x-samchon-controller":"AppController"}},"/new-arrivals":{"get":{"tags":["new-arrivals"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["new_arrivals","findAll"],"x-samchon-controller":"NewArrivalController"}},"/hero":{"get":{"tags":["hero"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["hero","findAll"],"x-samchon-controller":"HeroController"}},"/health":{"get":{"tags":["health"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["health","check"],"x-samchon-controller":"HealthController"}},"/gift-card/redeem":{"post":{"tags":["gift-cards"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"additionalProperties":false,"properties":{"code":{"type":"string"}},"required":["code"],"type":"object"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["gift_card","redeem"],"x-samchon-controller":"GiftCardController"}},"/gift-card/balance":{"get":{"tags":["gift-cards"],"parameters":[{"name":"code","in":"query","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["gift_card","balance"],"x-samchon-controller":"GiftCardController"}},"/gift-card":{"get":{"tags":["gift-cards"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["gift_card","findAll"],"x-samchon-controller":"GiftCardController"},"post":{"tags":["gift-cards"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGiftCardDto"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["gift_card","create"],"x-samchon-controller":"GiftCardController"}},"/gift-card/{id}":{"get":{"tags":["gift-cards"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["gift_card","findOne"],"x-samchon-controller":"GiftCardController"},"patch":{"tags":["gift-cards"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGiftCardDto"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["gift_card","update"],"x-samchon-controller":"GiftCardController"},"delete":{"tags":["gift-cards"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"responses":{"204":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["gift_card","remove"],"x-samchon-controller":"GiftCardController"}},"/coupon/validate":{"post":{"tags":["coupons"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"additionalProperties":false,"properties":{"code":{"type":"string"},"orderValue":{"type":"number"},"userId":{"type":"number"}},"required":["code","orderValue"],"type":"object"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["coupon","validate"],"x-samchon-controller":"CouponController"}},"/coupon":{"get":{"tags":["coupons"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["coupon","findAll"],"x-samchon-controller":"CouponController"},"post":{"tags":["coupons"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCouponDto"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["coupon","create"],"x-samchon-controller":"CouponController"}},"/coupon/analytics":{"get":{"tags":["coupons"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["coupon","analytics","getAnalytics"],"x-samchon-controller":"CouponController"}},"/coupon/{id}":{"patch":{"tags":["coupons"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCouponDto"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["coupon","update"],"x-samchon-controller":"CouponController"},"delete":{"tags":["coupons"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"responses":{"204":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["coupon","remove"],"x-samchon-controller":"CouponController"}},"/cart":{"post":{"tags":["cart"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"additionalProperties":false,"properties":{"productId":{"type":"number"},"quantity":{"type":"number"}},"required":["productId"],"type":"object"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["cart","create"],"x-samchon-controller":"CartController"},"get":{"tags":["cart"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["cart","findAll"],"x-samchon-controller":"CartController"}},"/cart/{id}":{"delete":{"tags":["cart"],"parameters":[{"name":"id","in":"path","schema":{"type":"string"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":null}}}}},"x-samchon-accessor":["cart","remove"],"x-samchon-controller":"CartController"}}},"components":{"schemas":{"CreateGiftCardDto":{"additionalProperties":false,"properties":{"code":{"type":"string"},"amount":{"type":"number"},"currency":{"type":"string"},"isActive":{"type":"boolean"},"expiresAt":{"oneOf":[{"format":"date-time","type":"string"},{"type":"string"}]}},"required":["code","amount"],"type":"object"},"UpdateGiftCardDto":{"additionalProperties":false,"properties":{"isActive":{"type":"boolean"},"expiredAt":{"oneOf":[{"format":"date-time","type":"string"},{"type":"string"}]},"expiresAt":{"oneOf":[{"format":"date-time","type":"string"},{"type":"string"}]}},"required":[],"type":"object"},"CreateCouponDto":{"additionalProperties":false,"properties":{"code":{"type":"string"},"discountType":{"oneOf":[{"const":"FIXED","description":null,"title":null},{"const":"PERCENTAGE","description":null,"title":null}]},"discountValue":{"type":"number"},"minOrderValue":{"type":"number"},"maxDiscount":{"type":"number"},"usageLimit":{"type":"number"},"perUserLimit":{"type":"number"},"validFrom":{"type":"string"},"validUntil":{"type":"string"}},"required":["code","discountType","discountValue","validFrom","validUntil"],"type":"object"}},"securitySchemes":{"bearer":{"type":"apiKey","name":"Authorization","in":"header"},"cookie":{"type":"apiKey","name":"accessToken","in":"cookie"}}},"tags":[{"name":"app"},{"name":"new-arrivals"},{"name":"hero"},{"name":"health"},{"name":"gift-cards"},{"name":"coupons"},{"name":"cart"}],"x-typia-emended-v12":true}